server refactor into untils and aux
