+++ python/dist-mcp/alation_ai_agent_mcp/__init__.py
@@ -1,3 +1,3 @@
-from .server import run_server
+from .server import run_server, create_server
 
-__all__ = ["run_server"]
+__all__ = ["run_server", "create_server"]
